dynmic meta data and favicon added
